tests/contract/test_database_schema.py
#!/usr/bin/env python3
"""Contract test for database schema.

This test validates that the database schema can be created and meets
all contract requirements. It MUST FAIL initially as no implementation exists.
"""

import pytest
import psycopg2
from psycopg2.extensions import ISOLATION_LEVEL_READ_COMMITTED
from typing import Dict, Any


class TestDatabaseSchemaContract:
    """Test database schema contract compliance."""

    def test_database_connection_requirements(self):
        """Test database connection requirements from contract."""
        # This will fail until database module is implemented
        from dedupe.config.database import get_connection_pool
        
        # Contract: Connection pooling with max 10 connections
        pool = get_connection_pool()
        assert pool.maxconn == 10
        
        # Contract: Connection timeout 30 seconds
        assert pool.timeout == 30

    def test_schema_creation(self):
        """Test that schema can be created per contract."""
        # This will fail until schema SQL is implemented
        from dedupe.config.database import create_schema
        
        # Should create all tables without errors
        create_schema()

    def test_tarball_record_table_structure(self):
        """Test TarballRecord table structure per contract."""
        # This will fail until models are implemented
        from dedupe.models.tarball_record import TarballRecord
        
        # Contract: Required fields with correct types
        record = TarballRecord(
            filename="test.tar.gz",
            hostname="server01",
            file_size=1024,
            status="PROCESSING"
        )
        
        # Contract: UUID primary key
        assert hasattr(record, 'id')
        assert record.status in ['PROCESSING', 'SUCCESS', 'PARTIAL', 'FAILED']

    def test_file_record_table_structure(self):
        """Test FileRecord table structure per contract."""
        # This will fail until models are implemented
        from dedupe.models.file_record import FileRecord
        
        # Contract: Required fields with constraints
        record = FileRecord(
            tarball_id="550e8400-e29b-41d4-a716-446655440000",
            filename="app.log",
            file_size=1024,
            checksum="abc123def456",
            hash_algorithm="sha256"
        )
        
        # Contract: File size must be non-negative
        assert record.file_size >= 0
        assert record.hash_algorithm in ['sha256', 'sha1', 'md5']

    def test_duplicate_group_table_structure(self):
        """Test DuplicateGroup table structure per contract."""
        # This will fail until models are implemented
        from dedupe.models.duplicate_group import DuplicateGroup
        
        # Contract: Unique checksum constraint
        group = DuplicateGroup(
            checksum="abc123def456",
            hash_algorithm="sha256",
            file_count=2
        )
        
        # Contract: Non-negative counters
        assert group.file_count >= 0
        assert group.total_size_saved >= 0

    def test_processing_log_table_structure(self):
        """Test ProcessingLog table structure per contract."""
        # This will fail until models are implemented
        from dedupe.models.processing_log import ProcessingLog
        
        # Contract: Required fields with constraints
        log = ProcessingLog(
            operation_type="PROCESS",
            log_level="INFO",
            message="Test message"
        )
        
        # Contract: Valid log levels
        assert log.log_level in ['DEBUG', 'INFO', 'WARNING', 'ERROR']

    def test_performance_indexes_exist(self):
        """Test that performance indexes are created per contract."""
        # This will fail until schema implementation exists
        from dedupe.config.database import get_connection
        
        conn = get_connection()
        cursor = conn.cursor()
        
        # Contract: Required performance indexes
        required_indexes = [
            'idx_filerecord_checksum',
            'idx_duplicategroup_checksum', 
            'idx_filerecord_tarball_id',
            'idx_tarballrecord_hostname',
            'idx_tarballrecord_processed_at',
            'idx_filerecord_filename',
            'idx_filerecord_duplicate_lookup',
            'idx_processinglog_operation_time'
        ]
        
        cursor.execute("""
            SELECT indexname FROM pg_indexes 
            WHERE tablename IN ('tarballrecord', 'filerecord', 'duplicategroup', 'processinglog')
        """)
        existing_indexes = [row[0] for row in cursor.fetchall()]
        
        for index in required_indexes:
            assert index in existing_indexes, f"Missing required index: {index}"

    def test_foreign_key_constraints(self):
        """Test that foreign key constraints are enforced per contract."""
        # This will fail until schema implementation exists
        from dedupe.config.database import get_connection
        
        conn = get_connection()
        cursor = conn.cursor()
        
        # Contract: FileRecord.tarball_id references TarballRecord.id
        cursor.execute("""
            SELECT COUNT(*) FROM information_schema.table_constraints 
            WHERE constraint_type = 'FOREIGN KEY' 
            AND table_name = 'filerecord'
            AND constraint_name LIKE '%tarball_id%'
        """)
        assert cursor.fetchone()[0] > 0

    def test_data_integrity_requirements(self):
        """Test data integrity requirements per contract."""
        # This will fail until implementation exists
        from dedupe.config.database import get_connection
        
        conn = get_connection()
        
        # Contract: Transaction isolation level READ_COMMITTED
        assert conn.isolation_level == ISOLATION_LEVEL_READ_COMMITTED
        
        # Contract: Query timeout 300 seconds
        cursor = conn.cursor()
        cursor.execute("SHOW statement_timeout")
        timeout = cursor.fetchone()[0]
        assert timeout == '300s' or timeout == '5min'

    def test_duplicate_detection_performance(self):
        """Test duplicate detection query performance per contract."""
        # This will fail until services are implemented
        from dedupe.services.duplicate_service import DuplicateService
        
        service = DuplicateService()
        
        # Contract: Query must complete in <2 seconds
        import time
        start = time.time()
        
        duplicates = service.find_duplicates_by_checksum("abc123def456")
        
        elapsed = time.time() - start
        assert elapsed < 2.0, f"Query took {elapsed}s, must be <2s per contract"